Ensure propagation of HTTP request errors.
When an error occurs in the core HTTP operations, an error event is emitted by a request object. For `putFile`, `putStream`, `readFile`, etc. the caller does not get a reference to the request object and does not get a chance to catch these errors  except through a global error handler. This change ensures HTTP request errors are propagated through the callback function.