Make `putStream` put any stream (finally). Closes #72.
Also reimplement `putFile` in terms of it, so that it streams files from disk. This loses the MD5 sadly, but it's a small price to pay.

Updated README with an example of a HTTP stream instead of another file stream, and added a test to make sure that example actually works.

Related: #14, #32, #48, #57.