Remove default canned ACL when access grants are used
Knox by default adds the canned 'public-read' ACL to PUT and COPY requests. However, if an app manually specifies an access grant such as 'x-amz-grant-read', the request will fail with the following error:

'Specifying both Canned ACLs and Header Grants is not allowed'

This behavior is documented on the respective API pages, these are to be used exclusive of each other.

This change simply removes the default x-amz-acl header when any of the access grant headers are specified by the app.