More auth signing tests; closes #6.
This contains all of the tests from #6 except for the problematic one with multiple headers. We'll wait for someone to actually want that before adding it.

There is a minor functionality improvement: x-amz-date headers are now ignored, instead of breaking the request, and if you use the auth library directly, you can pass strings instead of `Date` instances for the `date` option.